+++ lib/kitchen/directions/bake_composite_chapters.rb
@@ -4,7 +4,7 @@ module Kitchen
   module Directions
     module BakeCompositeChapters
       def self.v1(book:)
-        book.search("[data-type='composite-chapter']").each do |chapter|
+        book.composite_chapters.each do |chapter|
           chapter.first("[data-type='document-title']").id =
             "composite-chapter-#{chapter.count_in(:book)}"
         end