+++ lib/kitchen/directions/move_solutions_to_answer_key/strategies/uphysics.rb
@@ -18,7 +18,7 @@ module Kitchen::Directions::MoveSolutionsToAnswerKey
       def bake_section(chapter:, append_to:, klass:)
         section_solutions_set = []
         chapter.search(".#{klass}").each do |section|
-          section.search('[data-type="solution"]').each do |solution|
+          section.search('div[data-type="solution"]').each do |solution|
             section_solutions_set.push(solution.cut)
           end
         end
@@ -31,7 +31,7 @@ module Kitchen::Directions::MoveSolutionsToAnswerKey
 
       def bake_from_notes(chapter:, append_to:, klass:)
         solutions = []
-        chapter.notes(".#{klass}").each do |note|
+        chapter.notes("$.#{klass}").each do |note|
           note.exercises.each do |exercise|
             solution = exercise.solution
             solutions.push(solution.cut) if solution